ride made its own command similar to vanilla 1.19.4
